Split out the OSM parking mapper into a separate app, temporarily WITHOUT mouseover.
