+++ Cargo.toml
@@ -13,6 +13,7 @@ members = [
   "map_editor",
   "map_gui",
   "map_model",
+  "parking_mapper",
   "sim",
   "tests",
   "traffic_seitan",
