+++ experiment/src/main.rs
@@ -2,7 +2,6 @@ mod game;
 
 fn main() {
     widgetry::run(widgetry::Settings::new("experiment"), |ctx| {
-        ctx.canvas.cam_zoom = 10.0; // TODO
         let app = map_gui::SimpleApp::new(ctx, abstutil::CmdArgs::new());
         let states = vec![game::Game::new(ctx, &app)];
         (app, states)
