+++ parking_mapper/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "parking_mapper"
+version = "0.1.0"
+authors = ["Dustin Carlino <dabreegster@gmail.com>"]
+edition = "2018"
+
+[features]
+default = ["map_gui/native", "reqwest", "widgetry/native-backend"]
+
+[dependencies]
+abstutil = { path = "../abstutil" }
+geom = { path = "../geom" }
+map_gui = { path = "../map_gui" }
+map_model = { path = "../map_model" }
+reqwest = { version = "0.10.8", optional = true, default-features=false, features=["blocking", "rustls-tls"] }
+widgetry = { path = "../widgetry" }
+xmltree = "0.10.1"
